Implemented internal tool to log inspiration actions and their corresponding reason
Summary: Our internal logging is incredibly important to the success of our products - we are able to make clear, data-driven decisions as a team thanks to these readings. However, given the incredible number of features we ship every cut, we risk the chance of messing this up. This internal tool allows employees to receive real-time updates on-device for all inspiration actions and reasons. All logs can be viewed in a clean, easy-to-use overlay that can be minimized and is draggable to prevent obscuring the testing flow.

Differential Revision: D5450237

fbshipit-source-id: 54383ad9ba65a683b455f12795acc8805f4381c1